% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards.R
\name{add_ards}
\alias{add_ards}
\title{Adds data to an Analysis Results Dataset}
\usage{
add_ards(
  data,
  statvars,
  statdesc = NULL,
  byvars = NULL,
  trtvar = NULL,
  paramcd = NULL,
  anal_var = NULL,
  anal_val = NULL
)
}
\arguments{
\item{data}{The input dataset to create analysis results for.}

\item{statvars}{A vector of column names that identify the desired results.
Statvar columns must be numeric.}

\item{statdesc}{A vector of column names that identify a description value
for each statvar.}

\item{byvars}{A vector of column names to use for by variables.}

\item{trtvar}{A column name to use for the treatment variable.}

\item{paramcd}{A description of the analysis parameter.}

\item{anal_var}{A column name for the analysis variable.}
}
\value{
The input data frame, unaltered.
}
\description{
Adds data to an Analysis Results Dataset
}
