% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{add_method}
\alias{add_method}
\title{Adds a method to a program module}
\usage{
add_method(module, name, func)
}
\arguments{
\item{module}{The module to add a method to.}

\item{name}{The name of the method.}

\item{func}{A function that contains the method to add.}
}
\description{
Adds a method to a program module
}
