% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{add_parameter}
\alias{add_parameter}
\title{Adds a parameter to a program module}
\usage{
add_parameter(
  module,
  name,
  value = NULL,
  data_type = NULL,
  input_type = NULL,
  label = NULL,
  options = NULL
)
}
\arguments{
\item{module}{The module to add a parameter to.}

\item{name}{The name to associate with the module.  This name should
be unique in the module cache.}

\item{value}{The default value to use for the parameter.}

\item{data_type}{The data type of the parameter value.  Should be specified
as an R data type class name.}

\item{input_type}{The type of input control to use for this parameter.}

\item{label}{The label for this parameter.}

\item{options}{A vector of options to use for parameters that accept multiple
options values.}
}
\description{
Adds a parameter to a program module
}
