\name{AgroData2}
\alias{AgroData2}
\docType{data}
\title{AgroData2}
\description{AgroData2}
\usage{data(AgroData2)}
\format{
  A data frame with 9 observations on the following 6 variables.
  \describe{
    \item{\code{Rgn}}{a factor with levels \code{2} \code{3}}
    \item{\code{PH}}{a numeric vector}
    \item{\code{SD}}{a numeric vector}
    \item{\code{PL}}{a numeric vector}
    \item{\code{GLWR}}{a numeric vector}
    \item{\code{GT}}{a numeric vector}
  }
}
\keyword{datasets}
