\name{CarbonStock}
\alias{CarbonStock}
\docType{data}
\title{CarbonStock}
\description{CarbonStock}
\usage{data(CarbonStock)}
\format{
  A data frame with 50 observations on the following 7 variables.
  \describe{
    \item{\code{CS}}{a numeric vector}
    \item{\code{SV}}{a numeric vector}
    \item{\code{LA}}{a numeric vector}
    \item{\code{SM}}{a numeric vector}
    \item{\code{TN}}{a numeric vector}
    \item{\code{AP}}{a numeric vector}
    \item{\code{CEC}}{a numeric vector}
  }
}
\keyword{datasets}
