\name{CochranBIB}
\alias{CochranBIB}
\docType{data}
\title{Balanced incomplete block design in corn}
\description{Balanced incomplete block design in corn}
\usage{data(CochranBIB)}
\format{
  A data frame with 52 observations on the following 3 variables.
  \describe{
    \item{\code{loc}}{a factor with levels \code{B01} \code{B02} \code{B03} \code{B04} \code{B05} \code{B06} \code{B07} \code{B08} \code{B09} \code{B10} \code{B11} \code{B12} \code{B13}}
    \item{\code{gen}}{a factor with levels \code{G01} \code{G02} \code{G03} \code{G04} \code{G05} \code{G06} \code{G07} \code{G08} \code{G09} \code{G10} \code{G11} \code{G12} \code{G13}}
    \item{\code{yield}}{a numeric vector}
  }
}
\keyword{datasets}
