% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{copy_module}
\alias{copy_module}
\title{Copy an existing module}
\usage{
copy_module(module, name, local_path = NULL)
}
\arguments{
\item{module}{The module to copy.  You can pass either a module object
or a module name.}

\item{name}{The name of the new module.}

\item{local_path}{The development directory for the new module. If the
directory does not exist, it will be created.}
}
\value{
Returns the copied module.
}
\description{
Function will copy the passed module and return the copy.
The copy will be located in the specified path, or copy it to the
local path if not specified.
}
\seealso{
Other module: 
\code{\link{add_parameter}()},
\code{\link{create_module}()},
\code{\link{module}()},
\code{\link{print.module}()},
\code{\link{print.parameter}()},
\code{\link{read_module}()},
\code{\link{run_module}()},
\code{\link{test_module}()},
\code{\link{write_module}()}
}
\concept{module}
