\name{Corr}
\alias{Corr}
\docType{data}
\title{Corr}
\description{Data for Correlation Analysis}
\usage{data(Corr)}
\format{
  A data frame with 48 observations on the following 4 variables.
  \describe{
    \item{\code{GY14}}{a numeric vector}
    \item{\code{N}}{a numeric vector}
    \item{\code{P}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
  }
}
\source{BEDDA}
\keyword{datasets}
