% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{create_batch}
\alias{create_batch}
\title{Create program batch}
\usage{
create_batch(lst, location, remote = NULL)
}
\arguments{
\item{lst}{A list of programs to create the batch for.  The parameter value
should be a list of program_parms objects.}

\item{location}{The location is a full or relative
path to a directory to copy the indicated programs.  In most cases, this
location will be the study program directory.}

\item{remote}{The location of the remote module cache.}
}
\value{
The path to the "run_all" file.
}
\description{
This function will copy a list of programs to a location and
create a batch file to run them.  The function accepts a list of program
parameter objects and a directory to copy the programs.
}
\seealso{
Other batch: 
\code{\link{program_parms}()},
\code{\link{run_program}()}
}
\concept{batch}
