% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{create_module}
\alias{create_module}
\title{Initialize a new module}
\usage{
create_module(
  name,
  local_path,
  description = "",
  major_version = 0L,
  minor_version = 0L,
  active = TRUE,
  template = FALSE,
  level = "dev",
  keywords = c(),
  dependancies = c()
)
}
\arguments{
\item{name}{The name of the new module.}

\item{local_path}{The development directory for the new module. If the
directory does not exist, it will be created.}

\item{description}{A description of the module.}

\item{major_version}{An integer that identifies the major version number.}

\item{minor_version}{An integer that identifies the minor version number.}

\item{active}{Whether the module should be activated or not.}

\item{template}{Whether the module should be marked as a template.}

\item{level}{The development level for this module.  Valid values are
"dev", "test", and "prod".}

\item{keywords}{A vector of keywords to use for module search functions.}

\item{dependancies}{A vector of packages on which this module is dependent.}
}
\description{
Function will create folder and template files to
begin development of a new \strong{star} module.
}
