\name{designAugmentedLSD}
\alias{designAugmentedLSD}
\alias{designAugmentedLSD.default}
\title{Augmented Design for Latin Square}
\description{
 Randomization for augmented design in Latin Square
}
\usage{
designAugmentedLSD(checkTrmt, newTrmt, trial = 1, numFieldRow, serpentine = FALSE, factorName = NULL, display = TRUE, file = NULL)
}
\arguments{
  \item{checkTrmt}{a}
  \item{newTrmt}{a}
  \item{trial}{a}
  \item{numFieldRow}{a}
  \item{serpentine}{a}
  \item{factorName}{a}
  \item{display}{a}
  \item{file}{a}
}
\author{
 Alaine A. Gulles
}
\keyword{ randomization }
