\name{designAugmentedRCB}
\alias{designAugmentedRCB}
\alias{designAugmentedRCB.default}
\title{Augmented Design for Randomized Complete Block}
\description{
  Randomization for augmented design in Randomized Complete Block
}
\usage{
designAugmentedRCB(checkTrmt, newTrmt, r = 2, trial = 1, numFieldRow = 2, serpentine = FALSE, factorName = NULL, display = TRUE, file = NULL)
}
\arguments{
  \item{checkTrmt}{a}
  \item{newTrmt}{a}
  \item{r}{a}
  \item{trial}{a}
  \item{numFieldRow}{a}
  \item{serpentine}{a}
  \item{factorName}{a}
  \item{display}{a}
  \item{file}{a}
}
\author{
 Alaine A. Gulles
}
\keyword{ randomization }
