\name{designRowColumn}
\alias{designRowColumn.default}
\alias{designRowColumn}
\title{design Row-Column}
\description{Randomization for Row-Column Design}
\usage{
designRowColumn(generate, r = 2, trial = 1, rowblkPerRep, rowPerRowblk, colblkPerRep, numFieldRow, serpentine = FALSE, file = NULL)
}
\arguments{
  \item{generate}{list}
  \item{r}{numeric}
  \item{trial}{numeric}
  \item{rowblkPerRep}{numeric}
  \item{rowPerRowblk}{numeric}
  \item{colblkPerRep}{numeric}
  \item{numFieldRow}{numeric}
  \item{serpentine}{logical}
  \item{file}{path}
}
\references{
  DiGGer package
}
\author{VIBartolome, AAGulles}
\keyword{design}
