\name{Dule_AlphaGxE}
\alias{Dule_AlphaGxE}
\docType{data}
\title{Dule_AlphaGxE}
\description{Dule_AlphaGxE}
\usage{data(Dule_AlphaGxE)}
\format{
  A data frame with 720 observations on the following 11 variables.
  \describe{
    \item{\code{Env}}{a factor with levels \code{A01WSCM} \code{A03WS-0} \code{A03WS-1} \code{AWS01-D} \code{UWS02-2} \code{UWS02-8}}
    \item{\code{Rep}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{Block}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
    \item{\code{Plot}}{a numeric vector}
    \item{\code{Geno}}{a factor with levels \code{AUS196} \code{AZUCENA} \code{B6144F-MR-6-0-0} \code{C22} \code{CT13370-12-2-M} \code{CT13377-4-2-M} \code{CT13382-8-3-M} \code{CT6510-24-1-2} \code{CT6516-24-3-2} \code{DINORADO} \code{IR47686-30-3-2} \code{IR55419-04} \code{IR55423-01} \code{IR60080-46A} \code{IR65261-09-1-B} \code{IR65907-116-1-B} \code{IR66417-18-1-1-1} \code{IR66421-062-1-1-2} \code{IR66424-1-2-1-5} \code{IR68702-072-1-4-B} \code{IR70358-84-1-1} \code{IR70360-38-1-B-1} \code{IR71524-44-1-1} \code{IR71525-19-1-1} \code{IR72768-15-1-1} \code{IRAT170} \code{IRAT177} \code{IRAT212} \code{IRAT216} \code{MARAVILHA} \code{PALAWAN} \code{PRIMAVERA} \code{UPLRI-5} \code{UPLRI-7} \code{VANDANA} \code{WAB181-18} \code{WAB56-125} \code{WAB638-1} \code{WAB96-1-1} \code{WAYRAREM}}
    \item{\code{FLOWER}}{a numeric vector}
    \item{\code{Duration}}{a numeric vector}
    \item{\code{PlantHt}}{a numeric vector}
    \item{\code{CropBF}}{a numeric vector}
    \item{\code{HI}}{a numeric vector}
    \item{\code{YLD}}{a numeric vector}
  }
}
\keyword{datasets}
