\name{Dune}
\alias{Dune}
\docType{data}
\title{Dune}
\description{Dune}
\usage{data(Dune)}
\format{
  A data frame with 20 observations on the following 31 variables.
  \describe{
    \item{\code{Species}}{a numeric vector}
    \item{\code{Belper}}{a numeric vector}
    \item{\code{Empnig}}{a numeric vector}
    \item{\code{Junbuf}}{a numeric vector}
    \item{\code{Junart}}{a numeric vector}
    \item{\code{Airpra}}{a numeric vector}
    \item{\code{Elepal}}{a numeric vector}
    \item{\code{Rumace}}{a numeric vector}
    \item{\code{Viclat}}{a numeric vector}
    \item{\code{Brarut}}{a numeric vector}
    \item{\code{Ranfla}}{a numeric vector}
    \item{\code{Cirarv}}{a numeric vector}
    \item{\code{Hyprad}}{a numeric vector}
    \item{\code{Leoaut}}{a numeric vector}
    \item{\code{Potpal}}{a numeric vector}
    \item{\code{Poapra}}{a numeric vector}
    \item{\code{Calcus}}{a numeric vector}
    \item{\code{Tripra}}{a numeric vector}
    \item{\code{Trirep}}{a numeric vector}
    \item{\code{Antodo}}{a numeric vector}
    \item{\code{Salrep}}{a numeric vector}
    \item{\code{Achmil}}{a numeric vector}
    \item{\code{Poatri}}{a numeric vector}
    \item{\code{Chealb}}{a numeric vector}
    \item{\code{Elyrep}}{a numeric vector}
    \item{\code{Sagpro}}{a numeric vector}
    \item{\code{Plalan}}{a numeric vector}
    \item{\code{Agrsto}}{a numeric vector}
    \item{\code{Lolper}}{a numeric vector}
    \item{\code{Alogen}}{a numeric vector}
    \item{\code{Brohor}}{a numeric vector}
  }
}
\keyword{datasets}
