\name{Fertilizer}
\alias{Fertilizer}
\docType{data}
\title{Fertilizer}
\description{Two Independent Sample}
\usage{data(Fertilizer)}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{\code{Fertilizer}}{a factor with levels \code{A} \code{B}}
    \item{\code{NumFruit}}{a numeric vector}
  }
}
\source{BEDDA}
\examples{
data(Fertilizer)
## maybe str(Fertilizer) ; plot(Fertilizer) ...
}
\keyword{datasets}
