% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{find_modules}
\alias{find_modules}
\title{Find Modules}
\usage{
find_modules(name = NULL, keywords = NULL, version = "latest")
}
\arguments{
\item{name}{The name of the module to search for.}

\item{keywords}{A vector of keywords to search for.}

\item{version}{A parameter that identifies the version or version to search for.
Valid values are "all", "latest", or a vector of versions in the form "vX.Y".}
}
\description{
Find Modules
}
