\name{GeneticDistMatrix}
\alias{GeneticDistMatrix}
\docType{data}
\title{fishData}
\description{fishData}
\usage{data(GeneticDistMatrix)}
\format{
  A data frame with 12 observations on the following 13 variables.
  \describe{
    \item{\code{X}}{a factor with levels \code{G001} \code{G002} \code{G003} \code{G004} \code{G005} \code{G006} \code{G007} \code{G008} \code{G009} \code{G010} \code{P1} \code{P2}}
    \item{\code{P1}}{a numeric vector}
    \item{\code{P2}}{a numeric vector}
    \item{\code{G001}}{a numeric vector}
    \item{\code{G002}}{a numeric vector}
    \item{\code{G003}}{a numeric vector}
    \item{\code{G004}}{a numeric vector}
    \item{\code{G005}}{a numeric vector}
    \item{\code{G006}}{a numeric vector}
    \item{\code{G007}}{a numeric vector}
    \item{\code{G008}}{a numeric vector}
    \item{\code{G009}}{a numeric vector}
    \item{\code{G010}}{a numeric vector}
  }
}
\keyword{datasets}
