\name{GomezMaizeCross}
\alias{GomezMaizeCross}
\docType{data}
\title{Maize Cross}
\description{Maize Cross
}
\usage{data(GomezMaizeCross)}
\format{
  A data frame with 800 observations on the following 8 variables.
  \describe{
    \item{\code{KernelNumber}}{a numeric vector}
    \item{\code{KernelColor}}{a factor with levels \code{white} \code{yellow}}
    \item{\code{KernelType}}{a factor with levels \code{flint} \code{sweet}}
    \item{\code{KernelCategory}}{a factor with levels \code{white flint} \code{white sweet} \code{yellow flint} \code{yellow sweet}}
    \item{\code{YellowKernel}}{a numeric vector}
    \item{\code{WhiteKernel}}{a numeric vector}
    \item{\code{FlintType}}{a numeric vector}
    \item{\code{SweetType}}{a numeric vector}
  }
}
\keyword{datasets}
