\name{GomezSplit2RCBD}
\alias{GomezSplit2RCBD}
\docType{data}
\title{GomezSplit2RCBD}
\description{Data for Split-Split Plot Design in RCBD}
\usage{data(GomezSplit2RCBD)}
\format{
  A data frame with 135 observations on the following 5 variables.
  \describe{
    \item{\code{Rep}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{Nitrogen}}{a factor with levels \code{N1} \code{N2} \code{N3} \code{N4} \code{N5}}
    \item{\code{Management}}{a factor with levels \code{M1} \code{M2} \code{M3}}
    \item{\code{Variety}}{a factor with levels \code{V1} \code{V2} \code{V3}}
    \item{\code{GY}}{a numeric vector}
  }
}
\examples{
data(GomezSplit2RCBD)
## maybe str(GomezSplit2RCBD) ; plot(GomezSplit2RCBD) ...
}
\keyword{datasets}
