\name{GomezTwoCropSeasonRCB}
\alias{GomezTwoCropSeasonRCB}
\docType{data}
\title{GomezTwoCropSeasonRCB}
\description{GomezTwoCropSeasonRCB}
\usage{data(GomezTwoCropSeasonRCB)}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{\code{Season}}{a factor with levels \code{Dry} \code{Wet}}
    \item{\code{Rep}}{a factor with levels \code{I} \code{II} \code{III}}
    \item{\code{NitrogenRate}}{a factor with levels \code{0} \code{60} \code{90} \code{120} \code{150}}
    \item{\code{GrainYield}}{a numeric vector}
  }
}
\keyword{datasets}
