\name{grp_pss}
\alias{grp_pss}
\docType{data}
\title{grp_pss}
\description{grp_pss}
\usage{data(grp_pss)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{VARIETY}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{REP}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{YIELD}}{a numeric vector}
  }
}
\keyword{datasets}
