\name{Heatmap}
\alias{Heatmap}
\title{Heatmap}
\description{
function for single environment analysis; outputs heatmap
}
\usage{
Heatmap(R, data, genAs = "fixed", row, column, respvar, model, env)
}
\arguments{
  \item{R}{a}
  \item{data}{a}
  \item{genAs}{a}
  \item{row}{a}
  \item{column}{a}
  \item{respvar}{a}
  \item{model}{a}
  \item{env}{a}
}
\author{Alaine A. Gulles}
\keyword{graph}
