\name{IASRIBIBD}
\alias{IASRIBIBD}
\docType{data}
\title{Balanced Incomplete Block Design}
\description{Balanced Incomplete Block Design}
\usage{data(IASRIBIBD)}
\format{
  A data frame with 28 observations on the following 3 variables.
  \describe{
    \item{\code{BLK}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7}}
    \item{\code{TRT}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7}}
    \item{\code{CALORIES}}{a numeric vector}
  }
}
\keyword{datasets}
