\name{IASRICorrReg}
\alias{IASRICorrReg}
\docType{data}
\title{IASRICorrReg}
\description{IASRICorrReg}
\usage{data(IASRICorrReg)}
\format{
  A data frame with 46 observations on the following 5 variables.
  \describe{
    \item{\code{sampleNum}}{a numeric vector}
    \item{\code{PP}}{a numeric vector}
    \item{\code{PH}}{a numeric vector}
    \item{\code{NGL}}{a numeric vector}
    \item{\code{Yield}}{a numeric vector}
  }
}
\keyword{datasets}
