\name{IASRIpca}
\alias{IASRIpca}
\docType{data}
\title{Principal Component}
\description{Principal Component}
\usage{data(IASRIpca)}
\format{
  A data frame with 46 observations on the following 5 variables.
  \describe{
    \item{\code{sn}}{a numeric vector}
    \item{\code{pp}}{a numeric vector}
    \item{\code{ph}}{a numeric vector}
    \item{\code{ngl}}{a numeric vector}
    \item{\code{yld}}{a numeric vector}
  }
}
\keyword{datasets}
