\name{JohnAlpha}
\alias{JohnAlpha}
\docType{data}
\title{JohnAlpha}
\description{JohnAlpha}
\usage{data(JohnAlpha)}
\format{
  A data frame with 72 observations on the following 5 variables.
  \describe{
    \item{\code{plot}}{a numeric vector}
    \item{\code{rep}}{a factor with levels \code{R1} \code{R2} \code{R3}}
    \item{\code{block}}{a factor with levels \code{B1} \code{B2} \code{B3} \code{B4} \code{B5} \code{B6}}
    \item{\code{gen}}{a factor with levels \code{G01} \code{G02} \code{G03} \code{G04} \code{G05} \code{G06} \code{G07} \code{G08} \code{G09} \code{G10} \code{G11} \code{G12} \code{G13} \code{G14} \code{G15} \code{G16} \code{G17} \code{G18} \code{G19} \code{G20} \code{G21} \code{G22} \code{G23} \code{G24}}
    \item{\code{yield}}{a numeric vector}
  }
}
\keyword{datasets}
