\name{KemptonRowCol}
\alias{KemptonRowCol}
\docType{data}
\title{KemptonRowCol}
\description{KemptonRowCol}
\usage{data(KemptonRowCol)}
\format{
  A data frame with 68 observations on the following 6 variables.
  \describe{
    \item{\code{rep}}{a factor with levels \code{R1} \code{R2}}
    \item{\code{row}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
    \item{\code{col}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7}}
    \item{\code{gen}}{a factor with levels \code{G01} \code{G02} \code{G03} \code{G04} \code{G05} \code{G06} \code{G07} \code{G08} \code{G09} \code{G10} \code{G11} \code{G12} \code{G13} \code{G14} \code{G15} \code{G16} \code{G17} \code{G18} \code{G19} \code{G20} \code{G21} \code{G22} \code{G23} \code{G24} \code{G25} \code{G26} \code{G27} \code{G28} \code{G29} \code{G30} \code{G31} \code{G32} \code{G33} \code{G34} \code{G35}}
    \item{\code{yield}}{a numeric vector}
  }
}
\keyword{datasets}
