\name{Maize}
\alias{Maize}
\docType{data}
\title{Maize}
\description{description1}
\usage{data(Maize)}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{ROW}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{COLUMN}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{MAIZE}}{a factor with levels \code{A} \code{B} \code{C} \code{D}}
    \item{\code{GRAINYIELD}}{a numeric vector}
  }
}
\examples{
data(Maize)
}
\keyword{datasets}
