\name{ManovaData1}
\alias{ManovaData1}
\docType{data}
\title{ManovaData1}
\description{ManovaData1}
\usage{data(ManovaData1)}
\format{
  A data frame with 16 observations on the following 5 variables.
  \describe{
    \item{\code{REP}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{TRT}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{YLD}}{a numeric vector}
    \item{\code{PHT}}{a numeric vector}
    \item{\code{PAN}}{a numeric vector}
  }
}
\keyword{datasets}
