% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\docType{data}
\name{modules}
\alias{modules}
\title{A cache of available report modules.}
\format{
An object of class \code{environment} of length 1.
}
\usage{
modules
}
\description{
A cache of available report modules.
}
\keyword{datasets}
