\name{OP_pss}
\alias{OP_pss}
\docType{data}
\title{OP_pss}
\description{OP_pss}
\usage{data(OP_pss)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{NITROGEN}}{a factor with levels \code{0} \code{60} \code{90} \code{120}}
    \item{\code{REP}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{YIELD}}{a numeric vector}
  }
}
\keyword{datasets}
