\name{PGTYPIC}
\alias{PGTYPIC}
\docType{data}
\title{PGTYPIC}
\description{PGTYPIC}
\usage{data(PGTYPIC)}
\format{
  A data frame with 32 observations on the following 6 variables.
  \describe{
    \item{\code{Block}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{Geno}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8}}
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
  }
}
\keyword{datasets}
