% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{print.module}
\alias{print.module}
\title{Prints information about module}
\usage{
\method{print}{module}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Follow on parameters to the print function.}

\item{verbose}{Whether to print the module in verbose mode.  Valid values
are TRUE and FALSE.}
}
\description{
Prints information about module
}
