% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{pull_module}
\alias{pull_module}
\title{Pull a module down from the cache}
\usage{
pull_module(mod, location = NULL)
}
\arguments{
\item{mod}{The module to pull.}

\item{location}{The directory in which to pull the specified module.}
}
\description{
Pull a module down from the cache
}
