% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{push_module}
\alias{push_module}
\title{Function to promote a module}
\usage{
push_module(module, level = "d")
}
\arguments{
\item{module}{The module to push or the location of the
development directory of the module to push.}

\item{level}{The status of the function.  Valid values are d, t, and p.}
}
\value{
The function will return the pushed module.
}
\description{
The \code{push_module} function promotes a module to the
remote module cache. The function accepts a module object and a promotion
level.
}
\seealso{
Other cache: 
\code{\link{find_modules}()},
\code{\link{pull_module}()},
\code{\link{refresh_modules}()},
\code{\link{set_config}()}
}
\concept{cache}
