% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{read_module}
\alias{read_module}
\title{Reads a program module from the file system}
\usage{
read_module(location)
}
\arguments{
\item{location}{The location to read the module from.}
}
\value{
A module located at the path provided.
}
\description{
Reads a program module from the file system
}
