% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{refresh_modules}
\alias{refresh_modules}
\title{Refresh the module cache}
\usage{
refresh_modules(pth = NULL)
}
\arguments{
\item{pth}{An optional path to the cache.  If NULL, it will use the
path associated with the package.}
}
\description{
Refresh the module cache
}
