% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{run_module}
\alias{run_module}
\title{Run a module}
\usage{
run_module(module, ...)
}
\arguments{
\item{module}{The module to run.}

\item{...}{Parameters for the module.}
}
\value{
The function will return the environment that the code runs in.
}
\description{
The \code{run_module} function executes a module.  The function
will pass parameters via an environment, and execute the module code
in that environment.  The environment will then be returned from the function
so it can be examined by the calling program.
}
\seealso{
Other module: 
\code{\link{add_parameter}()},
\code{\link{copy_module}()},
\code{\link{create_module}()},
\code{\link{module}()},
\code{\link{print.module}()},
\code{\link{print.parameter}()},
\code{\link{read_module}()},
\code{\link{set_globals}()},
\code{\link{test_module}()},
\code{\link{write_module}()}
}
\concept{module}
