% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{run_program}
\alias{run_program}
\title{Run a program}
\usage{
run_program(program, ...)
}
\arguments{
\item{program}{The full or relative path of the program to run.}

\item{...}{Parameters for the module.}
}
\value{
The function will return the environment that the code runs in.
}
\description{
The \code{run_program} function executes a program.  The function
will pass parameters via an environment, and execute the program code
in that environment.  The environment will then be returned from the function
so it can be examined by the calling program.
}
\seealso{
Other batch: 
\code{\link{create_batch}()},
\code{\link{program_parms}()}
}
\concept{batch}
