% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{set_config}
\alias{set_config}
\title{Gets and sets the configuration file on the local machine}
\usage{
set_config(cache_directory = NULL, local_directory = NULL)
}
\arguments{
\item{cache_directory}{The path of the global module cache.}

\item{local_directory}{The path of the local module development area.}
}
\value{
The updated configuration.  Configuration is returned as a list.
}
\description{
The \code{set_config} function opens the configuration
file on the local machine and sets it with any passed parameters.  If
no parameters are passed, the function simply returns the existing config.
If no config file exists, it will be created.
}
\seealso{
Other cache: 
\code{\link{find_modules}()},
\code{\link{pull_module}()},
\code{\link{push_module}()},
\code{\link{refresh_modules}()}
}
\concept{cache}
