% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{set_globals}
\alias{set_globals}
\title{Gets and sets global information}
\usage{
set_globals(path = NULL, ...)
}
\arguments{
\item{path}{The name of the new globals file.  Should have a ".globals" extension.
Default is NULL, meaning the function will search for the nearest globals
file and return the object.}

\item{...}{Other parameters dynamically defined for this information file.}
}
\value{
The global information object.
}
\description{
A function to instantiate a new global information object.  The
object contains global information shared among programs.
The function creates, edits, or returns the global information file.
}
\seealso{
Other module: 
\code{\link{add_parameter}()},
\code{\link{copy_module}()},
\code{\link{create_module}()},
\code{\link{module}()},
\code{\link{print.module}()},
\code{\link{print.parameter}()},
\code{\link{read_module}()},
\code{\link{run_module}()},
\code{\link{test_module}()},
\code{\link{write_module}()}
}
\concept{module}
