% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{setup_module}
\alias{setup_module}
\title{Initialize a new module}
\usage{
setup_module(name, location)
}
\arguments{
\item{name}{The name of the new module.}

\item{location}{The development directory for the new module. If the
directory does not exist, it will be created.}
}
\description{
Function will create folder and template files to
begin development of a new \strong{star} module.
}
