\name{soil_redox}
\alias{soil_redox}
\docType{data}
\title{soil_redox}
\description{soil_redox}
\usage{data(soil_redox)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{DAS}}{a numeric vector}
    \item{\code{EH}}{a numeric vector}
  }
}
\keyword{datasets}
