\name{ssa.pairwise}
\alias{ssa.pairwise}
\title{ssa.pairwise}
\description{
function for single environment analysis; performs pairwise mean comparisons
}
\usage{
ssa.pairwise(model, type = "Tukey", alpha = 0.05, control.level = NULL)
}
\arguments{
  \item{model}{a}
  \item{type}{a}
  \item{alpha}{a}
  \item{control.level}{a}
}
\author{Alaine A. Gulles}
\keyword{pairwise}
