\name{ssaTest}
\alias{ssaTest}
\title{Single Environment Analysis}
\description{
 Performs Single Environment Analysis
}
\usage{
ssaTest(exptl.design = c("RCB", "BIBD", "AugRCB", "AugLS", "Alpha", "RowCol"), data, respvar, geno, row, column = NULL, rep = NULL, env = NULL, is.random = FALSE, exclude = NULL, descriptive = FALSE, varComp = FALSE, pairwise = c("none", "Dunnett", "Tukey"), alpha = 0.05, diagPlot = FALSE, outputPath = NULL)
}
\arguments{
  \item{exptl.design}{a}
  \item{data}{a}
  \item{respvar}{a}
  \item{geno}{a}
  \item{row}{a}
  \item{column}{a}
  \item{rep}{a}
  \item{env}{a}
  \item{is.random}{a}
  \item{exclude}{a}
  \item{descriptive}{a}
  \item{varComp}{a}
  \item{pairwise}{a}
  \item{alpha}{a}
  \item{diagPlot}{a}
  \item{outputPath}{a}
}
\author{VIB and AAG}
\keyword{ single-site }
