% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\docType{package}
\name{star}
\alias{star}
\alias{star-package}
\title{A Framework for Modular Program Extensions}
\description{
The \strong{modular} package provides an extensible framework for sharing
code within an organization.

The functions contained in the \strong{star} package are as follows:
\itemize{
  \item{\code{\link{module}}: Creates a program module}
  \item{\code{\link{push_module}}: Pushes a module to the cache}
}
Note that...
}
