% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{star_push}
\alias{star_push}
\title{Function to promote a module}
\usage{
star_push(func, level = "d")
}
\arguments{
\item{func}{The module to push.}

\item{level}{The status of the function.  Valid values are d, t, and p.}
}
\description{
Function to promote a module
}
