% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{test_module}
\alias{test_module}
\title{Test a module}
\usage{
test_module(mod = NULL)
}
\arguments{
\item{mod}{The program module to test.}
}
\description{
Test a module
}
\seealso{
Other module: 
\code{\link{add_parameter}()},
\code{\link{copy_module}()},
\code{\link{create_module}()},
\code{\link{module}()},
\code{\link{print.module}()},
\code{\link{print.parameter}()},
\code{\link{read_module}()},
\code{\link{run_module}()},
\code{\link{set_globals}()},
\code{\link{write_module}()}
}
\concept{module}
