\name{ToWide}
\alias{ToWide}
\alias{ToWide.default}
\title{title1}
\description{description1}
\usage{
ToWide(data, vnames, timevar, idvar)
}
\arguments{
  \item{data}{data frame}
  \item{vnames}{character vector}
  \item{timevar}{character vector}
  \item{idvar}{character vector}
}
\author{AAGulles}
\keyword{data}
