\name{White_heads}
\alias{White_heads}
\docType{data}
\title{White_heads}
\description{White_heads}
\usage{data(White_heads)}
\format{
  A data frame with 42 observations on the following 3 variables.
  \describe{
    \item{\code{Variety}}{a factor with levels \code{BPI-76} \code{C168-134} \code{C4-63G} \code{IR1108-3-5} \code{IR127-80-1} \code{IR1561-228-3-3} \code{IR20-1} \code{IR2061-464-2} \code{IR5} \code{Luma-1} \code{MRC407-1} \code{PARC 2-2} \code{Rexoro} \code{TN1}}
    \item{\code{Rep}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{WhiteHeads}}{a numeric vector}
  }
}
\keyword{datasets}
