\name{Year}
\alias{Year}
\docType{data}
\title{Year}
\description{Year}
\usage{data(Year)}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{YEAR}}{a factor with levels \code{1995} \code{1996}}
    \item{\code{TREAT}}{a factor with levels \code{CONTROL} \code{CR} \code{FYM} \code{NP} \code{NP+FYM} \code{NPK}}
    \item{\code{REPS}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{GRNYLD}}{a numeric vector}
  }
}
\keyword{datasets}
