% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{add_parameter}
\alias{add_parameter}
\title{Adds a parameter to a program module}
\usage{
add_parameter(
  module,
  name,
  default = NULL,
  data_type = NULL,
  input_type = NULL,
  label = NULL,
  description = NULL,
  options = NULL
)
}
\arguments{
\item{module}{The module to add a parameter to.}

\item{name}{The name to associate with the module.  This name should
be unique in the module cache.}

\item{default}{The default value to use for the parameter.}

\item{data_type}{The data type of the parameter value.  Should be specified
as an R data type class name.}

\item{input_type}{The type of input control to use for this parameter.}

\item{label}{The label for this parameter.}

\item{description}{A brief description for this parameter.}

\item{options}{A vector of options to use for parameters that accept multiple
options values.}
}
\value{
The input module, with new parameter added.

The modified module with new parameter added to the module
parameter list.
}
\description{
The \code{add_parameter} function adds a parameter to a module.
Each module is initalized with a parameter list.  This function adds a
parameter to the list.  The arguments of the function define the new
parameter.
}
\seealso{
Other module: 
\code{\link{copy_module}()},
\code{\link{create_module}()},
\code{\link{module}()},
\code{\link{print.module}()},
\code{\link{print.parameter}()},
\code{\link{read_module}()},
\code{\link{run_module}()},
\code{\link{set_globals}()},
\code{\link{test_module}()},
\code{\link{write_module}()}
}
\concept{module}
