% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{copy_module}
\alias{copy_module}
\title{Copy an existing module}
\usage{
copy_module(module, name, local_path = NULL)
}
\arguments{
\item{module}{The module to copy.  You can pass either a module object
or a module name.}

\item{name}{The name of the new module.}

\item{local_path}{The development directory for the new module. If the
directory does not exist, it will be created.}
}
\description{
Function will create folder and template files to
begin development of a new \strong{star} module.
}
