% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards.R
\name{init_ards}
\alias{init_ards}
\title{Initialize the Analysis Results Dataset}
\usage{
init_ards(
  studyid = NULL,
  tableid = NULL,
  adsns = NULL,
  population = NULL,
  time = NULL,
  where = NULL,
  reset = TRUE
)
}
\arguments{
\item{studyid}{The study for which the analysis was performed. Optional string value.}

\item{tableid}{A table identifier to use for the results. Optional string
value.}

\item{adsns}{A vector of source dataset names.}

\item{population}{A description of the analysis population.}

\item{time}{A description of the time frame used in the analysis.}

\item{where}{A description of the criteria used to subset the data for analysis.}

\item{reset}{If true, clear out existing ards dataset and replace with
empty template.  Otherwise, just assign parameters to template.}
}
\description{
Initialize the Analysis Results Dataset
}
