% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module}
\alias{module}
\title{Defines a new program module}
\usage{
module(
  name,
  description = "",
  major_version = 0L,
  minor_version = 0L,
  TA = NULL,
  level = "dev",
  keywords = c(),
  dependancies = c()
)
}
\arguments{
\item{name}{The name of the new module.}

\item{description}{A description of the module.}

\item{major_version}{An integer that identifies the major version number.}

\item{minor_version}{An integer that identifies the minor version number.}

\item{TA}{The therapeutic area to assign this module to.}

\item{level}{The development level for this module.  Valid values are
"dev", "test", and "prod".}

\item{keywords}{A vector of keywords to use for module search functions.}

\item{dependancies}{A vector of packages on which this module is dependent.}
}
\description{
Defines a new program module
}
