% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{modules}
\alias{modules}
\title{A list of available report modules.}
\format{
An object of class \code{list} of length 0.
}
\usage{
modules
}
\description{
A list of available report modules.
}
\keyword{datasets}
