% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{print.module}
\alias{print.module}
\title{Prints information about module}
\usage{
\method{print}{module}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Follow on parameters to the print function.}

\item{verbose}{Whether to print the module in verbose mode.  Valid values
are TRUE and FALSE.}
}
\value{
The module, invisibly.
}
\description{
A function to print a module.  The print out
will include most attributes of the module, organized in a readable fashion.
The module parameters will also be included in the print out.
}
\seealso{
Other module: 
\code{\link{add_parameter}()},
\code{\link{copy_module}()},
\code{\link{create_module}()},
\code{\link{module}()},
\code{\link{print.parameter}()},
\code{\link{read_module}()},
\code{\link{run_module}()},
\code{\link{test_module}()},
\code{\link{write_module}()}
}
\concept{module}
