% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{print.parameter}
\alias{print.parameter}
\title{Prints information about parameter}
\usage{
\method{print}{parameter}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The parameter to print.}

\item{...}{Follow on parameters to the print function.}

\item{verbose}{Whether to print the module in verbose mode.  Valid values
are TRUE and FALSE.}
}
\value{
The parameter, invisibly.
}
\description{
A function to print a module parameter.  The print out
will include most attributes of the parameter, organized in a readable fashion.
}
\seealso{
Other module: 
\code{\link{add_parameter}()},
\code{\link{copy_module}()},
\code{\link{create_module}()},
\code{\link{module}()},
\code{\link{print.module}()},
\code{\link{read_module}()},
\code{\link{run_module}()},
\code{\link{test_module}()},
\code{\link{write_module}()}
}
\concept{module}
