% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{pull_module}
\alias{pull_module}
\title{Pull a module down from the cache}
\usage{
pull_module(name, version = NULL, location = NULL)
}
\arguments{
\item{name}{The module name to pull.}

\item{version}{The module version to pull.  By default, the function
will pull the latest version.}

\item{location}{The directory in which to pull the specified module.  By
default, the function will pull the module into a directory with the same
name as the module in the current working directory.}
}
\description{
Pull a module down from the cache
}
