% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{pull_module}
\alias{pull_module}
\title{Pull a module down from the cache}
\usage{
pull_module(name, version = NULL, location = NULL)
}
\arguments{
\item{name}{The module name to pull.}

\item{version}{The module version to pull.  By default, the function
will pull the latest version.}

\item{location}{The directory in which to pull the specified module.  By
default, the function will pull the module into a directory with the same
name as the module in the current working directory.}
}
\value{
The function will return the pulled module.  Any file associated
with the module will be copied to the local path.
}
\description{
The \code{pull_module} function pulls a module and associated
files down from the module cache and places them in a local directory.  This
function is used to edit a module or examine the code.  You may specify a
specific version of the module.  If  a version is not specified, the function
will return the most recent version.
}
\seealso{
Other cache: 
\code{\link{find_modules}()},
\code{\link{push_module}()},
\code{\link{refresh_modules}()},
\code{\link{set_config}()}
}
\concept{cache}
