% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{push_module}
\alias{push_module}
\title{Function to promote a module}
\usage{
push_module(module, level = "d")
}
\arguments{
\item{module}{The module to push or the location of the
development directory of the module to push.}

\item{level}{The status of the function.  Valid values are d, t, and p.}
}
\description{
Function to promote a module
}
