% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{read_module}
\alias{read_module}
\title{Reads a program module from the file system}
\usage{
read_module(location, version = NULL)
}
\arguments{
\item{location}{The location to read the module from.}

\item{version}{The version of the module to read.}
}
\value{
A module located at the path provided.

Returns the module at the specified path.
}
\description{
To read a module from disk, use the \code{read_module} function.
The function will read the module into a object and return it.
}
\seealso{
Other module: 
\code{\link{add_parameter}()},
\code{\link{copy_module}()},
\code{\link{create_module}()},
\code{\link{module}()},
\code{\link{print.module}()},
\code{\link{print.parameter}()},
\code{\link{run_module}()},
\code{\link{set_globals}()},
\code{\link{test_module}()},
\code{\link{write_module}()}
}
\concept{module}
