% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{refresh_modules}
\alias{refresh_modules}
\title{Refresh the module cache}
\usage{
refresh_modules(pth = NULL)
}
\arguments{
\item{pth}{An optional path to the cache.  If NULL, it will use the
path associated with the package.}
}
\value{
A dataframe of available modules.
}
\description{
Read the available modules from the module cache and creates
a data frame with information about the modules.
}
\seealso{
Other cache: 
\code{\link{find_modules}()},
\code{\link{pull_module}()},
\code{\link{push_module}()},
\code{\link{set_config}()}
}
\concept{cache}
