\name{STAR-package}
\alias{STAR-package}
\alias{STAR}
\docType{package}
\title{Statistical Tools for Agricultural Research}
\description{ 
  A package for data management and basic statistical analysis
  of experimental data
}
\details{
\tabular{ll}{
Package: \tab STAR\cr
Type: \tab Package\cr
Version: \tab 2.0.2\cr
Date: \tab 2014-09-03\cr
Depends: \tab R (>= 3.0.2) \cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Alaine A. Gulles <a.gulles@irri.org> and Christoffer Edd N. Relente <c.relente@irri.org>, with contributions from
Violeta A. Bartolome, Leilani A. Nora, Rose Imee Zhella A. Morantte, and Guoyou Ye of 
International Rice Research Institute

Maintainer: Alaine A. Gulles <a.gulles@irri.org> and Christoffer Edd N. Relente <c.relente@irri.org>
}
\keyword{ package }
