\name{STAR-package}
\alias{STAR-package}
\alias{STAR}
\docType{package}
\title{
Spike Train Analysis with R
}
\description{
Functions to analyze neuronal spike trains
}
\details{
\tabular{ll}{
Package: \tab STAR\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2007-11-07\cr
Depends: \tab survival, mgcv, R2HTML, sound\cr
Suggests: \tab gam, lattice, ggplot2, HiddenMarkov\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.biomedicale.univ-paris5.fr/physcerv/C\_Pouzat/STAR.html\cr
}
}
\author{
Christophe Pouzat

Maintainer: Christophe Pouzat <christophe.pouzat@gmail.com>
}
\keyword{ package }
