\name{STAR-package}
\alias{STAR-package}
\alias{STAR}
\docType{package}
\title{
Spike Train Analysis with R
}
\description{
Functions to analyze neuronal spike trains
}
\details{
\tabular{ll}{
Package: \tab STAR\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2008-11-02\cr
Depends: \tab gss, survival, mgcv, R2HTML\cr
Suggests: \tab lattice, HiddenMarkov, snow, rstream\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://sites.google.com/site/spiketrainanalysiswithr\cr
}
}
\author{
Christophe Pouzat

Maintainer: Christophe Pouzat <christophe.pouzat@gmail.com>
}
\keyword{ package }
