% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{write_module}
\alias{write_module}
\title{Writes a program module to the file system}
\usage{
write_module(mod, location = NULL)
}
\arguments{
\item{mod}{The module to write.}

\item{location}{The file system location to write the module to.}
}
\description{
Writes a program module to the file system
}
